KOGITO-2298: Renamed Process list related components
