+++ packages/common/src/components/Templates/KogitoPageLayout/KogitoPageLayout.css
