--- packages/management-console/src/components/Atoms/ProcessBulkModalComponent/ProcessBulkModalComponent.tsx
@@ -1,77 +1,0 @@
-import React from 'react';
-import {
-  Modal,
-  Title,
-  TitleLevel,
-  BaseSizes,
-  Button,
-  TextContent,
-  Text
-} from '@patternfly/react-core';
-import ProcessBulkListComponent from '../ProcessBulkListComponent/ProcessBulkListComponent';
-
-interface IOwnProps {
-  modalTitle: JSX.Element;
-  modalContent?: string;
-  handleModalToggle: () => void;
-  abortedMessageObj?: any;
-  completedMessageObj?: any;
-  isModalOpen: boolean;
-  checkedArray: string[];
-  isAbortModalOpen?: boolean;
-  isSingleAbort?: any;
-}
-const Modalbox: React.FC<IOwnProps> = ({
-  modalContent,
-  modalTitle,
-  abortedMessageObj,
-  completedMessageObj,
-  isModalOpen,
-  checkedArray,
-  handleModalToggle,
-  isAbortModalOpen,
-  isSingleAbort
-}) => {
-  return (
-    <Modal
-      isSmall={true}
-      title=""
-      header={
-        <Title headingLevel={TitleLevel.h1} size={BaseSizes['2xl']}>
-          {modalTitle}
-        </Title>
-      }
-      isOpen={isModalOpen}
-      onClose={handleModalToggle}
-      actions={[
-        <Button
-          key="confirm-selection"
-          variant="primary"
-          onClick={handleModalToggle}
-        >
-          OK
-        </Button>
-      ]}
-      isFooterLeftAligned={false}
-    >
-      {abortedMessageObj !== undefined &&
-        completedMessageObj !== undefined &&
-        isAbortModalOpen && (
-          <ProcessBulkListComponent
-            abortedMessageObj={abortedMessageObj}
-            completedMessageObj={completedMessageObj}
-            isSingleAbort={isSingleAbort}
-            checkedArray={checkedArray}
-            isAbortModalOpen={isAbortModalOpen}
-          />
-        )}
-      <TextContent>
-        <Text>
-          <strong>{modalContent}</strong>
-        </Text>
-      </TextContent>
-    </Modal>
-  );
-};
-
-export default Modalbox;
