+++ packages/management-console/src/components/Templates/ProcessDetailsPage/ProcessDetailsPage.tsx
@@ -30,7 +30,7 @@ import ProcessDetailsProcessVariables from '../../Organisms/ProcessDetailsProces
 import ProcessDetailsTimeline from '../../Organisms/ProcessDetailsTimeline/ProcessDetailsTimeline';
 import './ProcessDetailsPage.css';
 import PageTitleComponent from '../../Molecules/PageTitleComponent/PageTitleComponent';
-import ProcessBulkModalComponent from '../../Atoms/ProcessBulkModalComponent/ProcessBulkModalComponent';
+import ProcessListModal from '../../Atoms/ProcessListModal/ProcessListModal';
 import {
   handleAbort,
   setTitle,
@@ -154,7 +154,7 @@ const ProcessDetailsPage: React.FC<
       {!error ? (
         <>
           <PageSection variant="light">
-            <ProcessBulkModalComponent
+            <ProcessListModal
               isModalOpen={isAbortModalOpen}
               handleModalToggle={handleAbortModalToggle}
               checkedArray={data && [data.ProcessInstances[0].state]}
@@ -168,7 +168,7 @@ const ProcessDetailsPage: React.FC<
               completedMessageObj={{}}
               isAbortModalOpen={isAbortModalOpen}
             />
-            <ProcessBulkModalComponent
+            <ProcessListModal
               isModalOpen={isModalOpen(
                 modalTitle,
                 isSkipModalOpen,
